backend/requirements-deploy.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
librosa==0.10.2
soundfile==0.12.1
numpy==2.4.3
scipy==1.17.1
scikit-learn==1.5.2
huggingface_hub==0.23.0
supabase==2.28.2
python-dotenv==1.2.2
yt-dlp==2026.3.17
